trimmed files that could be packaged in serverless
