+++ src/util/interface.ts
@@ -1,0 +1,27 @@
+export interface ITag {
+    id: string;
+    name: string;
+    color: string;
+}
+
+export interface BlogArticle {
+    id: string;
+    url: string;
+    tags: ITag[];
+    modifiedDate: string;
+    publishDate?: string;
+    title: string;
+    description: string;
+    link?: string;
+}
+
+export interface Project {
+    id: string;
+    url: string;
+    tags: ITag[];
+    modifiedDate: string;
+    title: string;
+    description: string;
+    link: string;
+    lastUpdated: string;
+}
